+++ R/article_info.R
@@ -10,10 +10,12 @@
 #'   \item{i}{The URL of the article from which the information is retrieved.}
 #'   \item{article_type}{The classification of the article (e.g., editorial, review).}
 #'   \item{Received}{The date the article was received by the publisher.}
+#'   \item{Revised}{The date the article was confirmed as revised by the publisher.}
 #'   \item{Accepted}{The date the article was accepted for publication.}
 #'   \item{tat}{The turnaround time, calculated as the number of days between the received and accepted dates.}
 #'   \item{year}{The year in which the article was accepted for publication.}
 #'   \item{issue_type}{Indicates whether the article is part of a special issue.}
+#'   \item{open_peer_review}{Indictes if article peer review is publicly available or not}
 #' }
 #' @examples
 #' url<-c("https://www.mdpi.com/2073-4336/8/4/45","https://www.mdpi.com/2073-4336/11/3/39")
@@ -64,7 +66,11 @@ article_info <- function(vector,sleep=2,sample_size,show_progress=TRUE) {
       } else {
         ex_paper2<- ex_paper2}
       
+      open_review<-paper%>%
+        html_nodes(".download .button--color-inversed")%>%
+        html_text2()
       
+      open_peer_review <- ifelse(any(grepl("Review Reports", open_review, ignore.case = TRUE)), "Yes", "No")
       
       article_type<-paper%>% # Type of article
         html_nodes(".articletype")%>%
@@ -90,7 +96,7 @@ article_info <- function(vector,sleep=2,sample_size,show_progress=TRUE) {
       
     })
     
-    temp_df<-data.frame(i,ex_paper,ex_paper2,article_type)
+    temp_df<-data.frame(i,ex_paper,ex_paper2,article_type,open_peer_review)
     paper_data<-bind_rows(paper_data,temp_df)
     
     count<-count+1
@@ -101,10 +107,13 @@ article_info <- function(vector,sleep=2,sample_size,show_progress=TRUE) {
     }
     
   }
+  
   final_table<-paper_data%>%
     mutate(Received=gsub("/.*","",tolower(ex_paper)), #Extract received data time and transform into date
            Received=gsub(".*received:","",Received),
            Received=as.Date(Received,"%d %B %Y"))%>%
+    mutate(Revised=str_extract(ex_paper, "(?<=Revised: )[0-9]{1,2} \\w+ \\d{4}"),
+           Revised=as.Date(Revised,"%d %B %Y"))%>%
     mutate(Accepted=gsub(".*accepted:","",tolower(ex_paper)), #Extract accepted time data and transform into date
            Accepted=gsub("/.*","",Accepted),
            Accepted=as.Date(Accepted,"%d %B %Y"))%>%