+++ R/plot_articles.R
@@ -2,8 +2,8 @@
 #' 
 #' @param articles_info Output dataframe from function articles_info.
 #' @param journal A string with the name of the journal for graph title purposes
-#' @param type select "summary","issues", "tat" or "type" depending on desired graph
-#' @import magrittr ggplot2 dplyr lubridate stringr scales
+#' @param type select "summary","issues", "tat", "review" or "type" depending on desired graph
+#' @import magrittr ggplot2 dplyr lubridate stringr scales tidyr
 #' @export plot_articles
 #' @return A plot (class: \code{ggplot}) depicting the desired information obtained from \code{article_info}
 #' @examples
@@ -12,12 +12,45 @@
 
 plot_articles<-function(articles_info,journal,type){
   
-  valid_types <- c("tat", "summary", "issues", "type")
+  valid_types <- c("tat", "summary", "issues", "type", "review")
   if (!type %in% valid_types) {
     stop("Error: parameter type must be 'summary, 'issues', 'tat' or 'type'")
   }
   
   switch(type,
+         
+         "review" = {
+           
+           open_review<-articles_info%>%
+             summarize(.by=c(year, open_peer_review), n=n())%>%
+             filter(if_all(everything(), ~ !is.na(.)))%>% #alternative to drop_na()
+             complete(year, open_peer_review = c("Yes", "No"), fill = list(n = 0))
+           
+           
+           graph<- ggplot(open_review,aes(x=year,y=n,fill=open_peer_review))+
+             geom_col()+
+             scale_fill_manual(values = c("#dc0000b2","#3366CC"))+
+             theme_classic()+
+             theme(text=element_text(size=20),
+                   plot.title = element_text(size=14))+
+             theme(axis.line.x = element_line(color="white", linewidth =  2),
+                   axis.line.y = element_line(color="white", linewidth =  2),
+                   axis.text.x = element_text(color="white"),
+                   axis.text.y = element_text(color="white"),
+                   legend.position = "bottom",
+                   legend.background = element_rect(fill = "#272822"),
+                   text=element_text(size=16,color = "white"),
+                   panel.background = element_rect(fill = "#272822"),
+                   plot.background = element_rect(fill = "#272822"))+
+             labs(x="Year", y="Articles", title = paste0("MDPI ",str_to_title(journal)," - Proportion of open reviews"),
+                  fill="Review available?")+
+             scale_y_continuous(limits = c(0,NA))
+           
+           graph
+           
+           
+         },
+         
          
          "tat" = {
            
@@ -35,7 +68,6 @@ plot_articles<-function(articles_info,journal,type){
            graph<- ggplot(tat,aes(x=year_month,y=mean_tat))+
              annotate("segment", x = min(tat$year_month),xend = max(tat$year_month),y = last_tat, 
                yend = last_tat, colour = "red", alpha = 0.7,linetype = "dotted")+
-             #geom_segment(aes(x=min(year_month),xend = max(year_month),y=last_tat,yend=last_tat),colour="red",alpha=.7,linetype = "dotted")+
              scale_x_date(date_labels = "%Y",breaks = date_breaks("year"))+
              geom_point(shape=21,colour="white",fill="#daeaf5",size=3)+